Tease out shared components from watch_slcd
